moves api stuff out to api file
